reco-py/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
grpcio-health-checking==1.66.1
grpcio-reflection==1.66.1